requirements.txt
adafruit-circuitpython-ssd1306
psutil 
pillow
adafruit_ssd1306
RPi.GPIO